# docdoku-server/docdoku-server-web/src/main/webapp/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "v2.5",
    "title" : "DocdokuPLM API"
  },
  "basePath" : "/api",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/workspaces/{workspaceId}/parts" : {
      "get" : {
        "tags" : [ "parts", "workspaces" ],
        "summary" : "Get part revisions",
        "description" : "",
        "operationId" : "getPartRevisions",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "start",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "length",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      },
      "post" : {
        "tags" : [ "parts", "workspaces" ],
        "summary" : "Create new part",
        "description" : "",
        "operationId" : "createNewPart",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/PartRevisionDTO"
            }
          }
        }
      }
    },
    "/workspaces/{workspaceId}/parts/checkedout" : {
      "get" : {
        "tags" : [ "parts", "workspaces" ],
        "summary" : "Get checked out part revisions",
        "description" : "",
        "operationId" : "getCheckedOutPartRevisions",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/PartRevisionDTO"
              }
            }
          }
        }
      }
    },
    "/workspaces/{workspaceId}/parts/count" : {
      "get" : {
        "tags" : [ "parts", "workspaces" ],
        "summary" : "Count part revisions",
        "description" : "",
        "operationId" : "getTotalNumberOfParts",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CountDTO"
            }
          }
        }
      }
    },
    "/workspaces/{workspaceId}/parts/countCheckedOut" : {
      "get" : {
        "tags" : [ "parts", "workspaces" ],
        "summary" : "Count checked out part revisions",
        "description" : "",
        "operationId" : "getCheckedOutNumberOfItems",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CountDTO"
            }
          }
        }
      }
    },
    "/workspaces/{workspaceId}/parts/import" : {
      "get" : {
        "tags" : [ "parts", "workspaces" ],
        "summary" : "Get current imports",
        "description" : "",
        "operationId" : "getImports",
        "consumes" : [ "multipart/form-data" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ImportDTO"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "parts", "workspaces" ],
        "summary" : "Import part attributes from file",
        "description" : "",
        "operationId" : "importAttributes",
        "consumes" : [ "multipart/form-data" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "autoCheckout",
          "in" : "query",
          "required" : false,
          "type" : "boolean"
        }, {
          "name" : "autoCheckin",
          "in" : "query",
          "required" : false,
          "type" : "boolean"
        }, {
          "name" : "permissiveUpdate",
          "in" : "query",
          "required" : false,
          "type" : "boolean"
        }, {
          "name" : "revisionNote",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/workspaces/{workspaceId}/parts/import/{importId}" : {
      "get" : {
        "tags" : [ "parts", "workspaces" ],
        "summary" : "Get import",
        "description" : "",
        "operationId" : "getImport",
        "consumes" : [ "multipart/form-data" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "importId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ImportDTO"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "parts", "workspaces" ],
        "summary" : "Delete import",
        "description" : "",
        "operationId" : "deleteImport",
        "consumes" : [ "multipart/form-data" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "importId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/workspaces/{workspaceId}/parts/numbers" : {
      "get" : {
        "tags" : [ "parts", "workspaces" ],
        "summary" : "Search part numbers",
        "description" : "",
        "operationId" : "searchPartNumbers",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "q",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/LightPartMasterDTO"
              }
            }
          }
        }
      }
    },
    "/workspaces/{workspaceId}/parts/parts_last_iter" : {
      "get" : {
        "tags" : [ "parts", "workspaces" ],
        "summary" : "Search documents last iteration to link",
        "description" : "",
        "operationId" : "searchDocumentsLastIterationToLink",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "q",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "l",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/PartIterationDTO"
              }
            }
          }
        }
      }
    },
    "/workspaces/{workspaceId}/parts/queries" : {
      "get" : {
        "tags" : [ "parts", "workspaces" ],
        "summary" : "Get custom queries",
        "description" : "",
        "operationId" : "getCustomQueries",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/QueryDTO"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "parts", "workspaces" ],
        "summary" : "Run custom queries",
        "description" : "",
        "operationId" : "runCustomQuery",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json", "application/octet-stream" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "save",
          "in" : "query",
          "required" : false,
          "type" : "boolean"
        }, {
          "name" : "export",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/QueryResult"
              }
            }
          }
        }
      }
    },
    "/workspaces/{workspaceId}/parts/queries/{queryId}" : {
      "delete" : {
        "tags" : [ "parts", "workspaces" ],
        "summary" : "Delete custom query",
        "description" : "",
        "operationId" : "deleteQuery",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "queryId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "CountDTO" : {
      "type" : "object",
      "properties" : {
        "count" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
...
    "ImportDTO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "fileName" : {
          "type" : "string"
        },
        "endDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "startDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "succeed" : {
          "type" : "boolean",
          "default" : false
        },
        "pending" : {
          "type" : "boolean",
          "default" : false
        },
        "errors" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "warnings" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        }
      }
    },
...
    "LightPartMasterDTO" : {
      "type" : "object",
      "properties" : {
        "partNumber" : {
          "type" : "string"
        },
        "partName" : {
          "type" : "string"
        }
      }
    },
...
    "QueryResult" : {
      "type" : "object",
      "properties" : {
        "query" : {
          "$ref" : "#/definitions/Query"
        },
        "rows" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/QueryResultRow"
          }
        },
        "exportType" : {
          "type" : "string",
          "enum" : [ "JSON", "CSV", "XLS" ]
        }
      }
    },
...
  }
}